pyproject.toml
[tool.poetry]
name = "ds-interview"
version = "0.1.0"
description = ""
authors = ["Lohith-reddy <sailohithgummi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.13"
streamlit = "^1.29.0"
langchain = "^0.0.352"
python-dotenv = "^1.0.0"
pypdf = "^3.17.3"
PyPDF2 = "^3.0.1"
google-generativeai = "^0.3.2"
openai = "^1.6.0"
chromadb = "^0.4.21"
google-cloud-aiplatform = "^1.38.1"
langchain-google-genai = "^0.0.5"
faiss-cpu = "^1.7.4"
qdrant-client = "^1.7.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"
black = "^23.12.0"
tiktoken = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120